Let LLVM codegen to handle multiple expressions
Previously, the expression compiler for LLVM was handling a single
expression at a time. With this change, the expression compiler now can
handle multiple expressions, it will also cache the computed results to
improve the performance.

Handling multiple expressions at a time is important because it
identifies matching CC list of the given character.