Implement `advance` and `match star` operations
The operations are used to match the input to the given RegEx.
The advance operation is used to forward the bit in the bit stream and
match star can match the given CC that repeated 0 or more times.

Introduce Basis bit streams as a dynamic `bitset` to store arbitrary
length of bits.
The CC and Marker bit streams are used to match CC and markers.
Their implementations inherit `BasisBitStream` and are identical, in
future they can be replaced directly by `BasisBitStream`. For now, they
make easier to understand the workflow.